pyproject.toml
[project]
name = "gameside"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
    "django>=6.0",
    "django-colorfield>=0.14.0",
    "pillow>=12.0.0",
]

[dependency-groups]
dev = [
    "django-browser-reload>=1.21.0",
    "factory-boy>=3.3.3",
    "ipython>=9.8.0",
    "pytest-django>=4.11.1",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "main.settings"
addopts = "-p no:warnings"
